stactools_pipelines/pipelines/noaa_hrrr/requirements.txt
aws-lambda-powertools
jmespath
requests
stactools-noaa-hrrr
s3fs
boto3
